flask_app/requirements.txt
Flask
pymongo
PyYAML
networkx
pyvis
dnspython